start on better api structure gen
