+++ .github/workflows/build_deploy_master_docs.yaml
@@ -54,7 +54,7 @@ jobs:
       - name: Install quarto
         uses: quarto-dev/quarto-actions/setup@v2
         with:
-          version: 1.2.335
+          version: 1.3.433
           tinytex: true
 
       - name: print quarto version
